requirements.txt
absl-py==2.2.2
attrs==25.3.0
blinker==1.9.0
cffi==1.17.1
click==8.2.0
colorama==0.4.6
comtypes==1.4.10
contourpy==1.3.2
cycler==0.12.1
Flask==3.1.0
flask-cors==5.0.1
flatbuffers==25.2.10
fonttools==4.58.0
itsdangerous==2.2.0
jax==0.6.0
jaxlib==0.6.0
Jinja2==3.1.6
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
mediapipe==0.10.20
ml_dtypes==0.5.1
numpy==1.26.4
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opt_einsum==3.4.0
packaging==25.0
pillow==11.2.1
protobuf==4.25.7
pycparser==2.22
pyparsing==3.2.3
# pypiwin32==223
python-dateutil==2.9.0.post0
pyttsx3==2.98
# pywin32==310
scipy==1.15.3
sentencepiece==0.2.0
six==1.17.0
sounddevice==0.5.1
Werkzeug==3.1.3
yt-dlp==2025.4.30
